move every component up one level
