+++ src/client/src/react/components/MainScore.css
