+++ src/client/src/react/components/MainScore.js
