+++ src/client/src/react/components/MainScore.test.js
