+++ src/client/src/react/components/Player.js
