+++ src/client/src/react/components/PlayerScore.js
