+++ src/client/src/react/components/main-score-state-func.js
