+++ src/client/src/react/pages/Login.js
@@ -1,7 +1,7 @@
 import React from "react";
 import "./Login.css";
 
-import ConnectionStatus from "../components/ConnectionStatus/ConnectionStatus";
+import ConnectionStatus from "../components/ConnectionStatus";
 import Title from "../components/Title";
 
 function Login({
